♻️ Run bundle when migrations present
Prior to this commit, we might have a situation where the initialize_app
container's gems are not what is indicated by the Gemfile.lock.  Which
creates problems when we run the `bundle exec rails db:create`; namely
the error will mention that we don't have the gem installed.

We can circumvent this via different docker compose files; however, it
seems reasonable to demand that we `bundle` before we execute `bundle
exec`; especially when we consider the fact that we're not always
running migrations.